+++ camphr/pipelines/knp/dependency_parser.py
