+++ pyproject.toml
@@ -110,7 +110,7 @@ transformers_sequence_classifier = "camphr.pipelines.transformers:TrfForSequence
 transformers_tokenizer = "camphr.pipelines.transformers:TrfTokenizer.from_nlp"
 juman_sentencizer = "camphr.pipelines.knp:juman_sentencizer_factory"
 knp = "camphr.pipelines.knp:KNP.from_nlp"
-knp_dependency_parser = "camphr.pipelines.knp_pipes.knp_dependency_parser:knp_dependency_parser_factory"
+knp_dependency_parser = "camphr.pipelines.knp.dependency_parser:knp_dependency_parser_factory"
 udify = "camphr.pipelines:Udify.from_nlp"
 elmo = "camphr.pipelines:Elmo.from_nlp"
 pattern_searcher = "camphr.pipelines:PatternSearcher"
