+++ tests/pipelines/knp/test_knp_dependency_parser.py
